requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
Flask-JWT-Extended
Werkzeug